+++ NEWS.md
@@ -10,6 +10,7 @@
   * Remove old files in final/ when make final
   * Check for correct texlive version
   * Check for empty notes/practical files
+  * Check for markdown versions of tufte
 
 # 0.5.0
   * Extract latex tags using xparse from latex